refactor(storage): own the FlutterSecureStorage provider
Move SecureStorageModule (the @module that registers FlutterSecureStorage
as a lazySingleton) from feature_auth's auth_local_data_source.dart into
packages/storage/lib/src/secure_storage_module.dart. The storage package
is the natural owner: it already depends on flutter_secure_storage and
its KeychainResetOnReinstall already consumed the registration. The
feature_auth DI module continues to consume FlutterSecureStorage via
gh<FlutterSecureStorage>(); only the registration site changes.
