+++ packages/features/auth/lib/src/data/datasources/auth_local_data_source.dart
@@ -124,18 +124,3 @@ class SecureStorageAuthDataSource implements AuthLocalDataSource {
     ]);
   }
 }
-
-@module
-abstract class SecureStorageModule {
-  @lazySingleton
-  FlutterSecureStorage provideSecureStorage() => const FlutterSecureStorage(
-    // iOS: tokens stay accessible after the first unlock (so background token
-    // refresh keeps working) but are device-only — never synced to iCloud
-    // Keychain and never restored onto a different device. Android needs no
-    // tuning here: this version already defaults to a strong AES-GCM +
-    // RSA-OAEP KeyStore backend.
-    iOptions: IOSOptions(
-      accessibility: KeychainAccessibility.first_unlock_this_device,
-    ),
-  );
-}
