+++ packages/features/auth/lib/src/di.module.dart
@@ -34,80 +34,55 @@ import 'package:network/network.dart' as _i372;
 import 'package:storage/storage.dart' as _i431;
 
 class FeatureAuthPackageModule extends _i526.MicroPackageModule {
-  // initializes the registration of main-scope dependencies inside of GetIt
+// initializes the registration of main-scope dependencies inside of GetIt
   @override
   _i687.FutureOr<void> init(_i526.GetItHelper gh) {
-    final secureStorageModule = _$SecureStorageModule();
     final authNetworkModule = _$AuthNetworkModule();
-    gh.lazySingleton<_i431.FlutterSecureStorage>(
-      () => secureStorageModule.provideSecureStorage(),
-    );
-    gh.lazySingleton<_i887.AuthLocalDataSource>(
-      () => _i887.SecureStorageAuthDataSource(gh<_i431.FlutterSecureStorage>()),
-    );
-    gh.lazySingleton<_i946.TokenRefresher>(
-      () => _i946.TokenRefresher(
-        gh<_i887.AuthLocalDataSource>(),
-        gh<_i372.Dio>(instanceName: 'plain'),
-      ),
-    );
-    gh.lazySingleton<_i372.Dio>(
-      () => authNetworkModule.provideDio(
-        gh<_i887.AuthLocalDataSource>(),
-        gh<_i946.TokenRefresher>(),
-        gh<_i259.EnvConfig>(),
-        gh<_i372.FirebasePerformance>(),
-      ),
-    );
+    gh.lazySingleton<_i887.AuthLocalDataSource>(() =>
+        _i887.SecureStorageAuthDataSource(gh<_i431.FlutterSecureStorage>()));
+    gh.lazySingleton<_i946.TokenRefresher>(() => _i946.TokenRefresher(
+          gh<_i887.AuthLocalDataSource>(),
+          gh<_i372.Dio>(instanceName: 'plain'),
+        ));
+    gh.lazySingleton<_i372.Dio>(() => authNetworkModule.provideDio(
+          gh<_i887.AuthLocalDataSource>(),
+          gh<_i946.TokenRefresher>(),
+          gh<_i259.EnvConfig>(),
+          gh<_i372.FirebasePerformance>(),
+        ));
     gh.lazySingleton<_i371.AuthRemoteDataSource>(
-      () => authNetworkModule.provideAuthRemoteDataSource(gh<_i372.Dio>()),
-    );
-    gh.lazySingleton<_i1063.AuthRepository>(
-      () => _i953.AuthRepositoryImpl(
-        gh<_i371.AuthRemoteDataSource>(),
-        gh<_i887.AuthLocalDataSource>(),
-        gh<_i946.TokenRefresher>(),
-      ),
-    );
+        () => authNetworkModule.provideAuthRemoteDataSource(gh<_i372.Dio>()));
+    gh.lazySingleton<_i1063.AuthRepository>(() => _i953.AuthRepositoryImpl(
+          gh<_i371.AuthRemoteDataSource>(),
+          gh<_i887.AuthLocalDataSource>(),
+          gh<_i946.TokenRefresher>(),
+        ));
     gh.factory<_i359.ChangePasswordUseCase>(
-      () => _i359.ChangePasswordUseCase(gh<_i1063.AuthRepository>()),
-    );
+        () => _i359.ChangePasswordUseCase(gh<_i1063.AuthRepository>()));
     gh.factory<_i884.DeleteAccountUseCase>(
-      () => _i884.DeleteAccountUseCase(gh<_i1063.AuthRepository>()),
-    );
+        () => _i884.DeleteAccountUseCase(gh<_i1063.AuthRepository>()));
     gh.factory<_i821.RegisterUseCase>(
-      () => _i821.RegisterUseCase(gh<_i1063.AuthRepository>()),
-    );
+        () => _i821.RegisterUseCase(gh<_i1063.AuthRepository>()));
     gh.factory<_i63.RestoreSessionUseCase>(
-      () => _i63.RestoreSessionUseCase(gh<_i1063.AuthRepository>()),
-    );
+        () => _i63.RestoreSessionUseCase(gh<_i1063.AuthRepository>()));
     gh.factory<_i147.SignInUseCase>(
-      () => _i147.SignInUseCase(gh<_i1063.AuthRepository>()),
-    );
+        () => _i147.SignInUseCase(gh<_i1063.AuthRepository>()));
     gh.factory<_i1002.SignOutUseCase>(
-      () => _i1002.SignOutUseCase(gh<_i1063.AuthRepository>()),
-    );
-    gh.factory<_i1061.DeleteAccountCubit>(
-      () => _i1061.DeleteAccountCubit(
-        gh<_i884.DeleteAccountUseCase>(),
-        gh<_i548.AnalyticsService>(),
-      ),
-    );
+        () => _i1002.SignOutUseCase(gh<_i1063.AuthRepository>()));
+    gh.factory<_i1061.DeleteAccountCubit>(() => _i1061.DeleteAccountCubit(
+          gh<_i884.DeleteAccountUseCase>(),
+          gh<_i548.AnalyticsService>(),
+        ));
     gh.factory<_i1062.ChangePasswordCubit>(
-      () => _i1062.ChangePasswordCubit(gh<_i359.ChangePasswordUseCase>()),
-    );
-    gh.lazySingleton<_i1014.AuthBloc>(
-      () => _i1014.AuthBloc(
-        signIn: gh<_i147.SignInUseCase>(),
-        register: gh<_i821.RegisterUseCase>(),
-        signOut: gh<_i1002.SignOutUseCase>(),
-        restoreSession: gh<_i63.RestoreSessionUseCase>(),
-        analytics: gh<_i548.AnalyticsService>(),
-      ),
-    );
+        () => _i1062.ChangePasswordCubit(gh<_i359.ChangePasswordUseCase>()));
+    gh.lazySingleton<_i1014.AuthBloc>(() => _i1014.AuthBloc(
+          signIn: gh<_i147.SignInUseCase>(),
+          register: gh<_i821.RegisterUseCase>(),
+          signOut: gh<_i1002.SignOutUseCase>(),
+          restoreSession: gh<_i63.RestoreSessionUseCase>(),
+          analytics: gh<_i548.AnalyticsService>(),
+        ));
   }
 }
-
-class _$SecureStorageModule extends _i887.SecureStorageModule {}
 
 class _$AuthNetworkModule extends _i692.AuthNetworkModule {}
