+++ packages/storage/lib/src/di.module.dart
@@ -9,24 +9,29 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart' as _i558;
 import 'package:injectable/injectable.dart' as _i526;
 import 'package:shared_preferences/shared_preferences.dart' as _i460;
 import 'package:storage/src/keychain_reset_on_reinstall.dart' as _i1049;
+import 'package:storage/src/secure_storage_module.dart' as _i837;
 import 'package:storage/src/shared_preferences_module.dart' as _i684;
 
 class StoragePackageModule extends _i526.MicroPackageModule {
-  // initializes the registration of main-scope dependencies inside of GetIt
+// initializes the registration of main-scope dependencies inside of GetIt
   @override
   _i687.FutureOr<void> init(_i526.GetItHelper gh) async {
     final sharedPreferencesModule = _$SharedPreferencesModule();
+    final secureStorageModule = _$SecureStorageModule();
     await gh.factoryAsync<_i460.SharedPreferences>(
       () => sharedPreferencesModule.provideSharedPreferences(),
       preResolve: true,
     );
+    gh.lazySingleton<_i558.FlutterSecureStorage>(
+        () => secureStorageModule.provideSecureStorage());
     gh.lazySingleton<_i1049.KeychainResetOnReinstall>(
-      () => _i1049.KeychainResetOnReinstall(
-        gh<_i460.SharedPreferences>(),
-        gh<_i558.FlutterSecureStorage>(),
-      ),
-    );
+        () => _i1049.KeychainResetOnReinstall(
+              gh<_i460.SharedPreferences>(),
+              gh<_i558.FlutterSecureStorage>(),
+            ));
   }
 }
 
 class _$SharedPreferencesModule extends _i684.SharedPreferencesModule {}
+
+class _$SecureStorageModule extends _i837.SecureStorageModule {}
